+++ wp-cloudflare-page-cache/src/Modules/Dashboard.php
@@ -285,6 +285,7 @@ class Dashboard implements Module_Interface {
 				'preloaderLocked'                     => ! Preloader_Process::can_start(),
 				'metrics'                             => Metrics::all(),
 				'ruleNeedsRepair'                     => get_option( Constants::KEY_RULE_UPDATE_FAILED ),
+				'advancedCacheWriteFailed'            => false !== get_option( Constants::KEY_ADVANCED_CACHE_WRITE_FAILED, false ),
 				'hasOverdueJobs'                      => $this->has_overdue_jobs(),
 				'invalidEncryptionState'              => Settings_Store::get_instance()->should_show_invalid_encryption_notice(),
 				'cloudflareConnected'                 => Settings_Store::get_instance()->is_cloudflare_connected(),